Move `convertTimestampPreview` into `web-common/lib`
